fix: 🧩 修复启动webpack时报错 (#230)
(cherry picked from commit 9a4b82564c1a3aa0c338ba7d5ab2f3e1660a7afc)
