+++ packages/@plugin/plugin-mobx/generator/index.cjs
@@ -2,8 +2,7 @@ module.exports = (generatorAPI) => {
   generatorAPI.extendPackage({
     devDependencies: {
       mobx: "^6.6.4",
-      "mobx-react-lite":"^3.2.2",
+      "mobx-react-lite": "^3.2.2",
     },
   });
 };
-
