+++ packages/core/template/template-webpack-script/generator/template/public/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Create-Neat</title>
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
+</html>
