+++ tsconfig.json
@@ -21,7 +21,6 @@
     "target": "ES5", // 编译成什么版本
     "module": "CommonJS", // 当前代码语法
     "declaration": false, // 是否自动创建类型声明文件
-    "allowJs": false, // 不允许编译javascript文件。
     "lib": [
       // 编译过程中需要引入的库文件的列表
       "es6",
